Library Data System
It is a system designed to manage library data recording processes. It helps us record and calculate processes such as book cataloging, borrowing, fee calculation, and returns. Additionally, it keeps these operations in memory, providing quick access to data whenever needed. It allows us to check the status of book inventory. This feature offers flexibility and ease of use for library staff.